Split into multiple files and tweak docs
