+++ docs/all.json
@@ -1,0 +1,1 @@
+{"shell":true,"revision":"397da4418f91e08eabd781903d39efef50d5a93c","objects":[{"type":null,"source":"lib/json_api_query_builder.ex:2","object_type":"ModuleObject","moduledoc":"Build Ecto queries from JSON-API requests.\n","module":"Elixir.JsonApiQueryBuilder","id":"JsonApiQueryBuilder"}],"language":"elixir","git_repo_url":"https://github.com/mbuhot/json_api_query_builder.git","client_version":"0.5.6","client_name":"inch_ex","branch_name":"test","args":[]}
